requirements.txt
streamlit
google-generativeai
mapbox
requests
python-dotenv
pandas
pydeck
folium
streamlit-folium
gspread
google-auth
